SONARJAVA-6827: Replace abstract class with utility class to avoid coupling
Move the shared visitor logic from AnonymousClassShouldBeLambdaCheck and
AnonymousClassOnFunctionalInterfaceCheck into a static method in
AnonymousClassToLambdaUtils. Both checks are now thin wrappers that
delegate to the utility, eliminating code duplication without introducing
an abstract class coupling.
