+++ java-checks/src/main/java/org/sonar/java/checks/AnonymousClassShouldBeLambdaCheck.java
@@ -16,25 +16,15 @@
  */
 package org.sonar.java.checks;
 
-import java.util.HashSet;
-import java.util.Set;
 import org.sonar.check.Rule;
 import org.sonar.java.checks.helpers.AnonymousClassToLambdaUtils;
 import org.sonar.plugins.java.api.JavaFileScanner;
 import org.sonar.plugins.java.api.JavaFileScannerContext;
 import org.sonar.plugins.java.api.JavaVersion;
 import org.sonar.plugins.java.api.JavaVersionAwareVisitor;
-import org.sonar.plugins.java.api.tree.BaseTreeVisitor;
-import org.sonar.plugins.java.api.tree.ClassTree;
-import org.sonar.plugins.java.api.tree.EnumConstantTree;
-import org.sonar.plugins.java.api.tree.IdentifierTree;
-import org.sonar.plugins.java.api.tree.NewClassTree;
 
 @Rule(key = "S1604")
-public class AnonymousClassShouldBeLambdaCheck extends BaseTreeVisitor implements JavaFileScanner, JavaVersionAwareVisitor {
-
-  private JavaFileScannerContext context;
-  private final Set<IdentifierTree> enumConstants = new HashSet<>();
+public class AnonymousClassShouldBeLambdaCheck implements JavaFileScanner, JavaVersionAwareVisitor {
 
   @Override
   public boolean isCompatibleWithJavaVersion(JavaVersion version) {
@@ -43,25 +33,7 @@ public class AnonymousClassShouldBeLambdaCheck extends BaseTreeVisitor implement
 
   @Override
   public void scanFile(JavaFileScannerContext context) {
-    this.context = context;
-    enumConstants.clear();
-    scan(context.getTree());
-  }
-
-  @Override
-  public void visitEnumConstant(EnumConstantTree tree) {
-    enumConstants.add(tree.simpleName());
-    super.visitEnumConstant(tree);
-    enumConstants.remove(tree.simpleName());
-  }
-
-  @Override
-  public void visitNewClass(NewClassTree tree) {
-    super.visitNewClass(tree);
-    ClassTree classBody = tree.classBody();
-    if (classBody != null && AnonymousClassToLambdaUtils.canBeConvertedToLambda(classBody, enumConstants)) {
-      context.reportIssue(this, tree.identifier(), "Make this anonymous inner class a lambda" + context.getJavaVersion().java8CompatibilityMessage());
-    }
+    AnonymousClassToLambdaUtils.scanForAnonymousClassConvertibleToLambda(this, context);
   }
 
 }
